requirements.txt
fastapi
sqlalchemy
alembic
psycopg2-binary
pytest
passlib
python-decouple 
python-jose 
python-multipart
pytz
sentry-sdk[fastapi]